--- src/MagicOnion.Client.Unity/Assets/Scripts/MagicOnion/MagicOnion.Client/MagicOnion.Shared/Utils.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4e8944691aa67234997771ba1a99d79f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
