+++ src/MagicOnion.Shared/Internal/FNV1A32.cs
@@ -1,6 +1,6 @@
-﻿using System.Text;
+using System.Text;
 
-namespace MagicOnion.Utils
+namespace MagicOnion.Internal
 {
     public static class FNV1A32
     {
