+++ tests/MagicOnion.Integration.Tests/HandCraftedStreamingHubClientTest.cs
@@ -1,9 +1,9 @@
 using System.Buffers;
 using Grpc.Net.Client;
 using MagicOnion.Client;
+using MagicOnion.Internal;
 using MagicOnion.Serialization;
 using MagicOnion.Server.Hubs;
-using MagicOnion.Utils;
 using MagicOnionTestServer;
 using Microsoft.AspNetCore.Mvc.Testing;
